Rename plugin to "Thread List", use proper thread sizes, cleanup
First, show as many threads as there are running on the system using OSCheckActiveThreads

Then, cleanup the NULL checks. They're good to have and some are overkill, but just in case, have them.

Third, rename to "Thread List". I feel like for the purpose of to a user it may be best to rename
this to "Process List" because that's what these are (I think), and it lines up with Rosalina's Process
List but the OS calls them "Threads", so, they are threads.